[INLONG-8154][Manager]Optimize the way of creating ExtractNodes, create ExtractNodeFactories to make it a single responsibility, easy to expand and maintain.
